backend/requirements.txt
pydantic-ai
google-genai
httpx
tavily-python
pydantic
python-dotenv
pytest
pytest-asyncio
fastapi
uvicorn
websockets
